refactor: Extract MobileHeader from ContactsListModal
no code change
